Split intermittent test into fail/pass paths
This tests has been intermittently failing for a while on CI. The reason
for this is because of the random generation of a content item based on
GovukSchemas gem.

The failing situation is caused when a change_note is created for a
content item and it has an update_type of major. The schema for
placeholder (and many others) don't support an array of content_history
in their details output for frontend or notification however these will
be provided by the Publishing API for any type that has change_notes.

This change splits this test into two similar ones which catch the
failing (change_history is present) and passing (change history not
present)