Clean up the content stores when importing content
The import endpoint deletes existing content with a matching content id.
Store the base paths in each content store before processing the import,
and afterwards. This information can then be used to work out what
content in the draft and live content stores needs deleting.

Add a test for this, and also update an existing test to stub the delete
requests.