Create concern for find_or_create_locked
This allows us to call `find_or_create_locked` on a model and allows the
same code to be used in multiple models.

The default value for locale has been dropped as we should have a
reasonable idea of locale by the point this has used, and having it as
optional encourages complacency over the field.