Catch CommandError when it is an error cause
In rails error matching involves checking the error and checking the
cause value of that error [1]. This breaks for us when we are expecting
an object of class CommandError and yet get a wrapped one.

This switches the value of `error` to be `error.cause` if error is not
an instance of `CommandError`. It'll still blow up if the cause is not a
`CommandError` but in that circumstance we have bigger problems.

[1]:
https://github.com/rails/rails/blob/ecc3a22bc3cd5e6de418ca40361418b468d19eea/activesupport/lib/active_support/rescuable.rb#L73-L94