Calculate total items using a window function.
Rather than run the expensive content_item_presenter query a second
time to get the total of all items, we can do it in the same query
using a window function over the whole result set.

This has involved a bit of refactoring in the GetContentCollection
class to make it possible to extract the total from the result rows.
The window function returns the (same) total in every row; we remove
the field from the presented results to avoid polluting them.