Specify a nil ordering so first doesn't order by id
We've had some problems where Postgres is picking different indexes to
use in different environments. Using the rails `.first` method on an
ActiveRecord scope creates an `ORDER BY` if one is not set on the query
which can cause the query to run slow in some environments.

This change removes the `order_by` which helps prevent the slow query.
Other options we have are: seeing if `VACUUM ANALYZE`, or even creating
an index that includes id for pillar based queries.