Split PutContent into separate concerns
CreateDraftContentItem creates a new draft content_item.
UpdateExistingDraftContentItem updates the existing one.
PutContent deals with actions that are similar to both cases.

It was quite difficult to see the differences between what happens when a
new content item was create vs an update, and this hopefully makes it
clearer what happens in each of those cases.