Add a migration to fix a whitehall Person object
This object was created and published, then deleted, then recreated
in whitehall.  Unfortunately it can't be published again because the
old object still lives in publishing-api with a different content_id
to the new object.  This migration updates the old object to share a
content_id with the new object and this allows the object to be published
from whitehall again.  We also fix the routes because the object is
a placeholder when published (routes not managed) and a gone when deleted
(routes are managed).