Remove all existing taxon tags
The education taxonomy is currently very alpha - it's not visible to the public
yet.

The finding things team has just completed a bulk tagging exercise with
publishers so we want to clear all existing tags and import the new ones.

This change removes those tags (links with type `taxons`).

There will be a follow up change to clear out remaining taxon content items,
that do not belong to the current taxonomy (i.e. there is no chain of
`parent_taxon` links linking them).