Remove data migrations that will no longer succeed
This removes the data migrations and their helpers that will no longer
be able to run with the changed code base. These were used to alter data
rather than the schema and have already been run in the production
environment.

On machines where these have been run there will be a **** NO FILE ****
shown on `rake db:migrate:status` for the timestamp of the migration.