WIP: allow lookup of redirected base paths
This request is used in content tagger and publishing apps to refer to pages
by base paths when those pages can be tagged to.

For example:
- related links
- topics
- taxons

The request currently silently ignores redirected editions.

For example,
/government/publications/impact-evaluation-of-the-send-pathfinder-programme was
unpublished and redirected to
/government/publications/send-pathfinder-programme-final-report

As far as the frontend is concerned, everything looks as expected, but I have
had trouble making sense of publishing api's record of this page and others
that have been redirected.

In this case, no non-superseded editions in publishing api are associated with
the *old URL*. The unpublished edition has the *new URL*. I haven't figured out
why this is, it could be intentional behaviour, or a historical bug (this
document was apparently unpublished multiple times).

I have had trouble understanding what the system is doing, because whitehall
also behaves in an unusual way: every unpublish is followed by the creation of
a new draft, using put content. This is so that an unpublished document can be
"un-unpublished" by going through 2i or force publish again.

Put content can sometimes have the side effect of creating separate content
items with document type "redirect". I originally thought that I should expect
to see one of these "redirect" editions for the *old URL*, which would mean
that lookup_by_base_path would be able to successfully look up the content id
for this page. **This does not appear to be the case**.

In fact, in this example, there is a "redirect" edition, but it's a draft. I
don't understand how this was created or why.

Regardless of how the data got into this state, I think the implementation of this
request is imprecise. I've made two changes to it:

1. Add a parameter to request additional information about base paths that have
been redirected. This allows to display a more useful error to the user and
suggest they tag to the redirect target instead.

I don't think this works at the moment, because it assumes that the unpublished
edition still has the old base path - which wasn't the case for the example I
looked at.

2. Don't return unpublished editions unless it was actually withdrawn (I
believe this was the original intent) Tagging to a "gone" or "not found" page
has no meaning, so I don't expect this to break anything.