Skip lookup of unpublished pages unless withdrawn
The current behaviour of this request returns the content ids for any base
paths that belong to published or unpublished pages.

The intention of including unpublished pages was to support withdrawn content:
these pages are still live, but show a message saying the content is not current.

However, there are 5 types of unpublishing, and the others mean content is not
visible on live.

It's not useful to complete the lookup in these cases. This request is used for
tagging, and tagging to content that isn't visible has no meaning.

There are also separate content items with document types of "redirect" and
"gone" which do not represent content. This means that the current
implementation is not very strict about only including content ids that are
"visible", and this makes it more difficult to predict what will and won't work
with this request.

We noticed this when looking into redirected content. Ideally we should be able
to distinguish content that has been redirected to another page (consolidated)
from content that never existed, so that publishers or users of content tagger
don't get confused (saying the content doesn't exist suggests they made a
mistake) and they can choose whether to tag to the new thing instead.

For now I think it's best to deliberately include this from the lookup, and
make an API change later to expose extra information about redirected content.
We are also considering whether we need to be able to query content ids
of base paths belonging to drafts

Trello:
https://trello.com/c/rC2btscW/448-content-tagger-show-appropriate-error-messages-when-uploading-untaggable-content
https://trello.com/c/ivoxhbhT/459-decide-what-we-need-from-lookup-by-base-path-so-that-we-can-handle-draft-and-redirected-content-when-tagging