Drop SchemaValidator double
As we use this test file for so many situations it is effectively an
integration test and the lack of actual schema validation seems to be
more of a problem than a benefit (nearly slipped through with Edition
links).

I think these tests here do probably need a rethink, there are very many
of them and we probably should separate out what is testing the class
and what is testing integration.