Add task for re-presenting content tagged to taxons
PR #745 fixed the link expansion between taxons, but it did not fix
those links in existing content. To fix this, we must re-present all
content pages which are tagged to taxons.

Running 'represent_downstream:all' is excessive in this case, so this
commit adds a new task which just re-presents affected content.