Refactored approach to Link Expansion
Since we are now doing recursion for both direct and reverse links this
is a refactoring that brings together both concepts into a single
LinkExpansion flow.

This introduces a number of new classes that have various
responsibilities in LinkExpansion:

LinkExpansion - The entry point to link expansion, and the only class
that's need to be interfaced with when performing link expansion.

LinkExpansion::Rules - A class that combines the rules of
Queries::DependeeExpansionRules and Queries::DependentExpansionRules
into a single class and is used as the source of knowledge as to what
links can be of depths.

LinkExpansion::LinkReference - A mediator that interfaces with queries
to load the links for link expansion.

LinkExpansion::ContentCache - A store of WebContentItems that are used
in populating links with content.

LinkGraph - A tree like structure to represent the graph of links that
are assembled as part of dependency resolution. It is not tied to
LinkExpansion so that it can be reused as part of dependency resolution.

LinkGraph::Node - An individual link within the graph. It can have
parent links and child links.

LinkGraph::NodeCollectionFactory - A class that interfaces with a
LinkReference to build the branches between LinkGraph::Nodes.