Support "search_in" parameter in get_content.
This allows specifying which fields to search against, as a comma
separated list. We support a single level of nesting with a dot
seperator. So, to search against title and the value inside the
description hash, do:

    ?search_in=title,description.value

If not provided, it will default to the existing title and base_path.