Handle edition links in LinkGraph
This updates the LinkGraph code to be aware of whether a link is the
result of an edition or a link set.

The LinkGraph::NodeCollectionFactory is aware that LinkReference has
edition links and link set links and calls the appropriate ones. Eg for
a root node links are an amalgamation of edition links and link set
whereas for a node it uses just the edition links or link set links
depending on what type of link the node is.

At the moment recursive link types are only supported with link set
links. Hopefully we don't reach a point where this complexity is also
needed in edition links.

This commit also replaces the initialisation attributes for LinkGraph
and LinkGraph::Node since there are now quite a lot of them a hash is
used instead.