Capture orphaned links before we delete them
Given a -> b, if we change a -> c, b still thinks it has a child of a,
but since we severed this relation, b has no reference to a, which means
that it is missed by dependency resolution.

This captures those relations before we remove them, and passes them
down into the dependency resolution worker.