Destroy inconsistent World Location News Articles
There are some inconsistencies in the way WLNA have been pushed to the
publishing api in the past, especially when translations have been
involved. A number of WLNA that have a `primary_locale` other than `en`
have documents in the publishing api with `en` as their locale. This
shouldn't actually happen. Deleting these instances and then
re-publishing from whitehall should fix the problem.