Index edition on (document_type, updated_at)
In get_linkables we check to see when the latest updated was to documents of a
particular type. This query can take some time to complete at the moment
because the document_type may not be updated very frequently, so the database
would have to scan through a lot of records if just using the updated_at index.

I've removed the plain document_type index as queries can use the first column
of a multicolumn index.
https://www.postgresql.org/docs/9.6/static/indexes-multicolumn.html

Use postgres's `CONCURRENTLY` option on the create index, because this can take
a while to create and we don't want to prevent concurrent inserts/updates
during this time.
https://robots.thoughtbot.com/how-to-create-postgres-indexes-concurrently-in