Compress memcache so we can store a bit more data
We have a problem that we're limited to 1MB of cache storage for
Memcache and we're trying to cache > 1MB of needs. This adds compression
that reduces our cache sizes to be below 1MB. It's not really an ideal
as if the payload increases this cache will fail but it can be done
without having to investigate any infrastructure changes.