Allow lookup requests to return draft content
We want to be able to tag to draft content in content tagger.

For example, content tagger is now the only way to tag certain things
to mainstream content (eg need ids).

This breaks at the moment because we can't lookup the content, which affects
both bulk tagging and the regular tagging workflow.

Although we need to handle draft/live content differently after the lookup,
I think this request should work with both.

Since we only pull out content ids we don't actually care which edition the
query uses for a given document, but I've ordered the states with published
first in case there are any edge cases where a base path is reused across
documents.