Update content_item_presenter benchmark
As this scope is used for the entire query it runs horrendously if you
provide it with no conditions to slim it down (as it has to scan every
single edition).

This has changed because previously this was used to collect a list of
items that would be included in an IN of the ids, so this behaviour has
been replicated within the scope query.