Update draft content store on Publish
Previously the publishing api was not updating the draft content store
when a piece of content was published with a change. This resulted in disparity
between the actual and expected content in the draft content store.

Specifically we noticed this because the publish command removes the
`access_limited` hash in the live content store but did not in the draft
content store.

We now will trigger teh `DownstreamDraftWorker` if there is a access
limited hash.

Mobbed with Christopher Baines and Emma Beynon