Simplify the "downstream" tests
These tests are failing because we've added extra data to the outgoing
payload. The testing of the exact payload here doesn't add any value,
because it's already tested in other parts of the system. By removing
them it's clearer what we're actually testing. The tests should be
refactored anyway, they're quite confusing at the moment.