Add v2 lookup-by-base-path request
This allows draft and unpublished documents to be looked up in addition
to live content.

See https://github.com/alphagov/publishing-api/issues/812

Locale and content id are returned for each successful lookup, as well as
unpublishing details if applicable.

I've also included document type, because there are some document types that
the application may not want. For example it makes no sense to tag something to
something with document type "redirect".