Return an empty link set response for new documents
PutContent used to create a LinkSet when creating a draft, but this got
removed in 050ca98

This affects `GET links` and `GET expanded-links`: both of these now
return 404 for documents that have never had links set.
This complicates building tagging interfaces that work with draft content.

LinkSet is an implementation detail: we expect to be able to get links
at any time regardless of publishing status of the document.