Fix `links` subquery
When an edition has links of more than one type the current sub-query in
`ContentItemPresenter` was raising as the sub-query returned multiple
rows.

This commit fixes the SQL to return the expected result by removing the
`group by` clause. The JSON aggregation already seems to do the
necessary grouping by type.