Restructure message queue test for better debugging
This fixes a number of cases where the message queue test fails because
the (valid) random content is not accepted by the publishing-api.

- Remove the special casing for the `change_note`. The random schema
generator will generate items with and without the change note, so I
think this is redundant. The message queue test is only meant to test
that the payload sent to the queue is correct.
- Add a nice error message when a call fails because of random content.
This makes it easier to debug the issue.