Remove Specialist Publisher parent links
Parent links were added to all specialist publisher documents in
https://github.com/alphagov/publishing-api/pull/821 . This has had an
unfortunate side effect - as parent links are represented as "children"
links, this caused finder content items such as /aaib-reports or
/dfid-research-outputs to have links to every child which was making the
content items huge in size.

The effects of this seem to be:
- 413 errors writing to content store
- slow responses from content store
- potentially (I hope this was the cause) a huge increase in memory
usage in content store.

This migration removes these parent links, which seems to be the best
immediate action we can take. As an intermediary message apps that plan
to consume this link can use the finder link which is the same as the
parent.