Include orphaned_content_ids in worker unique credentials
It was reported that our clearing of orphaned_content_ids was behaving
inconsistently. A reason this could be happening is that jobs to clear
them could be rejected as orphaned_content_ids do not appear in the
sidekiq uniq jobs criteria.

This adds them in so they will not be discarded.