Trigger end-to-end tests build
This is fired off from any build that gets past the rspec tests. We'll
have to learn whether this is annoying or not depending on how long it
takes to run and how backlogged the tests are.

Ideally I wanted to have it run just when we have a PR but that seemed
to need a really hacky solution.