Verify pacts against production content store
This seems a more cautious default path to take than running against
master. This would have caught the issue we had earlier where we could
not deploy to production as
https://github.com/alphagov/content-store/pull/272 was not deployed.

In situations where we want to get greens against merged branches we can
always build against master build with the parameter.

I wasn't sure if something similar should be done for GDS API Adapters
or not since that's a gem rather than something that's deployed :-/