Modify test for link_set/edition links conflict
We had an issue where dependency resolution was failing to update items
due to it only using edition links when both link set and edition links
are available.

The precedence behaviour makes sense when expanding links but not
when determining dependencies.

Issue was pointed out by Andrew Garner when he noticed a dependency had
not been updated with a Whitehall republishing.