Make LinkReference classes responsible for merging links
This has been moved out of the code shared for link_graphs into
the individual LinkReference classes for LinkExpansion and
DependencyResolution. This has been done because the merging is slightly
different in those cases: for dependency resolution you want all links
and you don't care if they are edition or link set, however for link
expansion you want link set links to take precedence.

This isn't exactly the prettiest solution as a whole bunch of arguments
are passed to links_by_link_type - I think we can't be too far from
refactoring this though as we want to explore setting links in schemas.