Use DistributedLock when writing to content store
This is to ensure that we don't make writes to the content store
concurrently. When this happens we can end up with the content store
being out of sync as it doesn't perform any locking.

You might wonder why this is being fixed here rather than in the Content
Store where it would probably make more sense. The reason that is done
here is to not contribute new complexity to the Content Store in line
with the goal that it can gradually become an S3 bucket of flat files.
So long as there is only one app writing to Content Store (this one)
it's fine to solve it here.