Add validation for redirect destination for a specific segments mode
When the segments mode is preserve, the query parameters from the
incoming url will be appended on to the destination. If the route type
is prefix, in addition to the query parameters, part of the incoming
path may also be appended.

This validation prevents the introduction of redirects where the
destination contains a fragment or where the destination has query
parameters, as its unlikely that these are appropriate to use when the
segments mode is preserve.