Populate change history for specialist docs without it
There are ~200,000 specialist documents which have an empty change history, this won't
get presented downstream because a memoizing clause in how the change history is presented.
So identify these records (they can be anything but draft), generate an initial change note
if none exists, then update the change history in the details hash from the change notes
for the edition.
This is a modification of the migration in https://github.com/alphagov/publishing-api/commit/c29d05b19dc2c89c21c9dabbaf29dbad2fa54373
it now finds all edition_ids for a given content id and retrieves change notes for them,
this gets round the earlier problem of duplicates being created.