Up sidekiq to a concurrency of 6
Currently we use ~95 connections of 200 on production, with less on
staging (seems a little strange) and less still on integration.

Publishing API currently uses 16 connections on prod/staging and 13 on
integration. This will increase it's usage to 28 on prod/staging and 21
on integration. Which still be well below the 200 connections limit.