Set last_edited_at to be previous items value
When a republish update type is used with PUT content it's possible to
lose the last_edited_at value, since this is semi automatically set by
the Publishing API (which is in itself a problem as the semantics of PUT
have gone a bit skewy). This is some defensive code that starts off by
setting last_edited_at to the previous value if one is available.