Deduplicate embedded specialist edition change history
This migration identifies editions with change history embedded in the details hash
which contain duplicate notes.
Some editions contain duplicate change history but are not associated with any change
notes, so we repopulate the change notes from the change history, deduplicating
based on document and note text.
Some of the duplicate change history may be legitimate entries so we clear
the entry in the details hash and represent the edition downstream.
The presenter will regenerate this history correctly based on the ChangeNote records.
These have been de-duplicated in an earlier migration.