Add timestamps to Documents
Documents should have timestamps so we know when they were created or updated.
We add timestamps and populate them primarily using Edition data, however there
are some cases where a Document has no editions, in which case we use Event
data instead. This is only possible for documents with an `en` locale only, so for
documents with multiple locales we use data from S3.