Remove rake task to import events for a `content_id`
This rake task was created to import all events relating to a
particular content_id. It does so by calculating the filename on S3
where the payload for a certain event was archived to.

This worked because we do archiving every week, so if a event happened
on Tuesday, we know that it will be archived into the CSV with a name
like `events/date-of-the-sunday-before.csv.gz`.

Because the job has been failing for a couple of weeks now, we can no
longer determine the correct file for each event.