Create exporter for a tagging event log
This creates a rake task that exports a CSV with a "tagging history".
It contains all the taxonomy-related tagging events. We use the CSV in
content tagger to show the history of how much was tagged to a taxon
when, and by whom. We use a CSV rather than an actual API to postpone a
decision on what the API should look like.

https://trello.com/c/GRPCKmbU