Allow world_location to be base_path-less
WorldLocation pages are going to become taxon based but we still need
them in the publishing API as content is linked to them in Whitehall and
we need this to continue for now.

So, as they will no longer be renderable we are making them base_path less.

This commit allows world_location to be base_path-less.