Add rake task to find redirected DFID reports
Add a temporary rake task which finds all published redirects with the
DFID base path (`/dfid-research-outputs/`). These were not removed from
the search API when the redirects were configured. This commit will
allow us to build a list of them in order to remove them from the search
API in bulk.

This rake task will be removed once it has been run in production.