Add missing `public_updated_at` dates for specialist-publisher docs
There are around 1146 published specialist-publisher docs with a missing
`public_updated_at` date. This prevents them from being republished as a
minor update because specialist-publisher requires a `public_updated_at`
date to generate a `public_timestamp` in the search index.

This migration populates that field with the best data available, which
is the `first_published_at` date. This is populated for all published
specialist-publisher documents.

There are long-term plans to simplifiy the document and edition
timestamps, so this is just a temporary workaround to allow all
specialist-publisher documents to be republished.