Allow lookup by base path filter criteria to be specified
In order to migrate content out of router-data we need 
to be able to find the content ID for unpublished
redirects. 

This change allow us to remove the filtering  required 
by mainstream applications for unpublished.type and 
document_type fields.

We have inverted the existing logic around unpublishing_type
filters based on the fact that content must be `published`
or `unpublished` and only `unpublished` content will
have an unpublishing record.