Restrict by_base_path to 'live' content
Restricting by only editions in the 'published' or 'unpublished'
states is not sufficient at the moment, as there are sometimes
multiple editions with published/unpublished states for a single base
path.

Currently, it's possible to get in to this problomatic situation when
an edition is published, where the previous edition is unpublished,
and the edition being published has a different content id to the
unpublished edition. The content store for the unpublished edition
will be NULL, but it will still be in the unpublished state.

Update the tests to set content_store to live.