Expand and rename LinkExpansion::Rules
This renamed LinkExpansion::Rules to be ExpansionRules so that it is no
longer under the namespace of LinkExpansion, which had become confusing
once we had specific classes for Rules::LinkExpansion and
Rules::DependencyResolution.

This change introduces a number of interrelated things:
- Fix use of un_reverse to unreverse.
- No longer has the concept of root_reverse_links, now just
reverse_links.
- Avoids the use of `reverse` as a keyword for when things are done the
opposite way round with dependency resolution - this meant a different
thing to reverse links. If needed the word `backwards` is used instead.
- Introduces LinkExpansion and DependencyResolution classes within the
ExpandedRules namespace for handling queries specific to their needs.
- Introduces MultiLevelLinks class within ExpandedRules for handling
concerns of working out the next links in mult level link groupings.
- Cache the paths of multi level links for the duration of the
application (since they're hardcoded).