Send first_published_at to GA when publishing
This will send a GA hit whenever an edition is published for the first time.
The information will be sent along as an `event`, using the `trackerId` currently in
use in govuk_admin_template (`tid: UA-26179049-6`).

The client id we're sending (`cid: 660ad712-9753-4cb9-97a7-c9e9f13c318e`) is
randomly generated and can be used to identify that the hits are coming from the
publishing-api rather than govuk_admin_template or govuk_frontend_toolkit (both
apps are normally used to send analytics information).