Cache preloaded data where nothing is found
When the preload doesn't find data for a content_id it previously stored
nothing in the cache so we would try look it up on demand.

This stores content_ids where nothing was found.

This gives us a small speed boost:

Given link expansion with a warm graph
```
> le = LinkExpansion.by_content_id("91b8ef20-74e7-4552-880c-50e6d73c2ff9")
> le.link_graph.to_h
``

Before:

```
> puts Benchmark.measure { le.instance_variable_set(:@content_cache, nil); le.links_with_content }
  0.320000   0.030000   0.350000 (  0.657331)
```

After:

```
> puts Benchmark.measure { le.instance_variable_set(:@content_cache, nil); le.links_with_content }
  0.300000   0.000000   0.300000 (  0.574295)
```