Return a 409 error upon failing to publish an edition with no draft.
The 400 error that this replaces is a poor semantic representation of the
nature of the problem encountered. From the
[MDN page](https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/400)

```
The HTTP 400 Bad Request response status code indicates that the server
could not understand the request due to invalid syntax. The client should
not repeat this request without modification.
```

A better fit seems to be the
[409 Conflict error](https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/409)

```
The HTTP 409 Conflict response status code indicates a request conflict
with current state of the server.
```