Get pagination self url by parsing URL
This means that the query parameters are correctly encoded in the
output rather than relying on the input to be. This is mainly a
problem in the pact tests where the self URL would be formated as
'fields[]=' and the next/previous URLs would be formatted as
'fields%5B%5D='.

We don't encounter this discrepancy in the offset/limit pagination
endpoints as they use regular expressions to modify the page=
parameter and leave all the others exactly as they came in.