Add payload_version to message queue messages
This is the same field passed to content store.

This allows the search indexing consumer to end up in the right state
when
publish/unpublish messages are processed out of order, because we can
ignore
version numbers that older than our existing copy for a given content
id.

This means for the requeue task, we'll have to query the events from the
database and pick the most recent one for each document. This could
probably be
written more efficiently but I'm expecting the bottleneck to be on the
consumer
side.