Fix editionless change notes
This migration makes a few different attempts to fix
or remove ChangeNote records which do not link to an Edition.

- If a ChangeNote has no edition associated with it
but there's a duplicate note on the document edition
sequence then delete this note.
- If a ChangeNote#note is "First published." and the
associated Document#editions have no "First published."
ChangeNote then assign this note to the first Edition in
the sequence.
- Failing to match any of the above, check the change_history
for the relevant document editions, cross reference the note
text and public_updated_at with the note text and public_timestamp
and associate the appropriate ChangeNote with the relevant Edition.
- Make another pass over ChangeNotes with edition:nil, if the number
of ChangeNotes for the given document matches the number of editions
for the same document then delete this note, it can't be reassigned.
- Lookup all ChangeNotes with a timestamp older than the earliest
edition for the same document, we only care about editions with no
change notes. Delete the older notes as they can't be associated
with an Edition.
- There's one ChangeNote left after this, so associated manually.

There are currently 1274 ChangeNotes with no edition associated
with them.