Change how last_edited_at is populated on Edition
In keeping with how the newer date fields are populated, we move
`last_edited_at` to be populated in the Edition::Timestamps model.  We also
make the decision to stop copying the `last_edited_at` value from the previous
edition in the case of a republish, as we feel it's simpler to just update it
at each PUT content call, and it's not a field that's used by the Content
Store.