Keep track of change history for links
This is the first PR to implement a change tracking feature for links.
It's half of https://github.com/alphagov/publishing-api/pull/1024.

We'll create a separate PR to expose the change history via the API.
That information will be used in content tagger by users who want to
see when a page was tagged to a certain taxon and who tagged what to
which taxon. This enables them to monitor tagging behaviour.

https://trello.com/c/162iqEul