Fix bulk link ordering
By using the join in bulk links we lost the ordering of links. This was
making the results returned unpredictable and vulnerable to failing
tests.

I chose to remove the join rather than adding ordering here as it
appears that the 2 queries run faster than a single one and it means we
don't need to re-state the default link ordering here.