Rename reverse links when looking up direct links
This fixes something that was overlooked when the link caching was
introduced. When a reverse has direct links, the resultant SQL query was
using the reverse link name and not finding results.

eg `WHERE link_type = 'child_taxons' AND other_link_type IN
('associated_taxons')`

Which would yield no results as child_taxons isn't in the database

This corrects this to be

`WHERE link_type = 'parent_taxons' AND other_link_type IN
('associated_taxons')`