Add endpoint for tagging history
This adds an endpoint to expose links history. It's a follow up to

The maximum number of results is limited to 250. This is an arbitrary
number to prevent too much data being returned, and to hold off from
implementing pagination. Once we have found a need for pagination it
can be implemented (using key-set pagination!).

The first consumer of this data will be content-tagger. We'll use it
there to show the history of content tagged to pages, and pages tagged
to certain taxons. In the future it might also be used in other
publisher applications like whitehall and mainstream publisher.

Trello: https://trello.com/c/Ays5PBSH